Streaming: Move BotFrameworkStreamingAdapter to botbuilder module
