--- libraries/botbuilder-streaming-extensions/.gitignore
@@ -1,4 +1,0 @@
-/lib
-/node_modules
-*.js.map
-/coverage/**
