--- libraries/botbuilder-streaming-extensions/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "botbuilder-streaming-extensions",
-  "version": "0.0.1",
-  "description": "",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.js",
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/chai": "^4.1.7",
-    "@types/node": "^10.12.18",
-    "chai": "^4.2.0",
-    "nyc": "^11.4.1",
-    "ts-node": "^4.1.0",
-    "tslint": "^5.16.0",
-    "tslint-microsoft-contrib": "^5.2.1",
-    "typescript": "3.1.1"
-  },
-  "dependencies": {
-    "@azure/ms-rest-js": "1.2.6",
-    "botbuilder": "~4.5.1",
-    "botbuilder-core": "~4.5.1",
-    "botframework-connector": "~4.5.1",
-    "botframework-schema": "~4.5.1",
-    "botframework-streaming": "~4.6.1",
-    "promise.prototype.finally": "^3.1.0",
-    "uuid": "^3.3.2",
-    "watershed": "^0.4.0"
-  },
-  "engines": {
-    "node": ">12.3"
-  },
-  "scripts": {
-    "test": "tsc && nyc mocha tests/",
-    "build": "tsc",
-    "clean": "erase /q /s .\\lib",
-    "set-version": "npm version --allow-same-version ${Version}"
-  },
-  "files": [
-    "/lib",
-    "/src"
-  ]
-}
