--- libraries/botbuilder-streaming-extensions/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "commonjs",
-    "declaration": true,
-    "sourceMap": true,
-    "outDir": "./lib",
-    "rootDir": "./src",
-    "types" : ["node"]
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
