+++ libraries/botframework-streaming/package.json
@@ -1,6 +1,6 @@
 {
   "name": "botframework-streaming",
-  "version": "4.6.1",
+  "version": "4.1.6",
   "description": "",
   "main": "lib/index.js",
   "typings": "lib/index.d.js",
